Split Rectangle to its own file
