+++ include/NAS2D/Renderer/Renderer.h
@@ -12,7 +12,7 @@
 
 #include "Color.h"
 #include "Point.h"
-#include "Primitives.h"
+#include "Rectangle.h"
 #include "../Signal.h"
 #include "../Resources/Image.h"
 #include "../Resources/Font.h"
